pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "relic"
version = "0.1.0"
description = "Codebase knowledge management CLI for AI coding agents"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1.40",
    "pyperclip>=1.8.2",
    "PyYAML>=6.0.1",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.scripts]
relic = "relic.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["relic"]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
